requirements.txt
click==8.1.3
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
SQLAlchemy==1.4.39
tqdm==4.64.0
Werkzeug==2.1.2
zipp==3.8.1
PyJWT~=2.4.0
pip~=22.2.1
wheel~=0.37.1
setuptools~=60.2.0